// src/shared/types/mock-server.ts
export type MockServerData = {
  id: number[];
  title: string[];
  description: string[];
  postDate: string[];
  city: string[];
  previewImage: string[];
  images: string[];
  isPremium: boolean[];
  isFavorite: boolean[];
  rating: number[];
  type: string[];
  bedrooms: number[];
  maxAdults: number[];
  price: number[];
  goods: string[];
  hostId: number[];
  hostName: string[];
  hostEmail: string[];
  hostAvatar: string[];
  hostPassword: string[];
  hostIsPro: boolean[];
  locationLat: number[];
  locationLng: number[];
  commentsCount: number[];
};